+++ ref-exchange/tests/common/utils.rs
@@ -182,6 +182,10 @@ pub fn get_deposits(
     view!(pool.get_deposits(account_id)).unwrap_json::<HashMap<String, U128>>()
 }
 
+pub fn list_referrals(pool: &ContractAccount<Exchange>) -> HashMap<String, u32> {
+    view!(pool.list_referrals(None, None)).unwrap_json::<HashMap<String, u32>>()
+}
+
 /// get ref-exchange's frozenlist tokens
 pub fn get_frozenlist(pool: &ContractAccount<Exchange>) -> Vec<String> {
     view!(pool.get_frozenlist_tokens()).unwrap_json::<Vec<String>>()
@@ -232,6 +236,15 @@ pub fn mft_balance_of(
     view!(pool.mft_balance_of(token_or_pool.to_string(), to_va(account_id.clone())))
         .unwrap_json::<U128>()
         .0
+}
+
+pub fn mft_has_registered(
+    pool: &ContractAccount<Exchange>,
+    token_or_pool: &str,
+    account_id: ValidAccountId,
+) -> bool {
+    view!(pool.mft_has_registered(token_or_pool.to_string(), account_id))
+        .unwrap_json::<bool>()
 }
 
 pub fn mft_total_supply(
@@ -309,7 +322,7 @@ pub fn setup_pool_with_liquidity() -> (
         contract_id: swap(),
         bytes: &EXCHANGE_WASM_BYTES,
         signer_account: root,
-        init_method: new(to_va("owner".to_string()), 4, 1)
+        init_method: new(to_va("owner".to_string()), 5, 0)
     );
     let token1 = test_token(&root, dai(), vec![swap()]);
     let token2 = test_token(&root, eth(), vec![swap()]);
@@ -410,7 +423,7 @@ pub fn setup_stable_pool_with_liquidity(
         contract_id: swap(),
         bytes: &EXCHANGE_WASM_BYTES,
         signer_account: root,
-        init_method: new(owner.valid_account_id(), 1600, 400)
+        init_method: new(owner.valid_account_id(), 2000, 0)
     );
 
     let mut token_contracts: Vec<ContractAccount<TestToken>> = vec![];
@@ -501,7 +514,7 @@ pub fn setup_rated_pool_with_liquidity(
         contract_id: swap(),
         bytes: &EXCHANGE_WASM_BYTES,
         signer_account: root,
-        init_method: new(owner.valid_account_id(), 1600, 400)
+        init_method: new(owner.valid_account_id(), 2000, 0)
     );
 
     let mut pool_tokens = vec![];
@@ -625,7 +638,7 @@ pub fn setup_rated_pool(
         contract_id: swap(),
         bytes: &EXCHANGE_WASM_BYTES,
         signer_account: root,
-        init_method: new(owner.valid_account_id(), 1600, 400)
+        init_method: new(owner.valid_account_id(), 2000, 0)
     );
 
     let mut pool_tokens = vec![];
@@ -750,7 +763,7 @@ pub fn mint_and_deposit_token(
     .assert_success();
 }
 
-pub fn setup_exchange(root: &UserAccount, exchange_fee: u32, referral_fee: u32) -> (
+pub fn setup_exchange(root: &UserAccount, admin_fee_bps: u32) -> (
     UserAccount,
     ContractAccount<Exchange>,
 ) {
@@ -760,7 +773,7 @@ pub fn setup_exchange(root: &UserAccount, exchange_fee: u32, referral_fee: u32) 
         contract_id: swap(),
         bytes: &EXCHANGE_WASM_BYTES,
         signer_account: root,
-        init_method: new(to_va("owner".to_string()), exchange_fee, referral_fee)
+        init_method: new(to_va("owner".to_string()), admin_fee_bps, 0)
     );
     (owner, pool)
 }