+++ ref-exchange/tests/test_migrate.rs
@@ -9,7 +9,7 @@ use crate::common::utils::*;
 pub mod common;
 
 near_sdk_sim::lazy_static_include::lazy_static_include_bytes! {
-    PREV_EXCHANGE_WASM_BYTES => "../res/ref_exchange.wasm",
+    PREV_EXCHANGE_WASM_BYTES => "../releases/ref_exchange_release_v162.wasm",
     EXCHANGE_WASM_BYTES => "../res/ref_exchange.wasm",
 }
 
@@ -47,17 +47,28 @@ fn test_upgrade() {
     .assert_success();
     let metadata = get_metadata(&pool);
     // println!("{:#?}", metadata);
-    assert_eq!(metadata.version, "1.6.2".to_string());
-    assert_eq!(metadata.exchange_fee, 4);
-    assert_eq!(metadata.referral_fee, 1);
+    assert_eq!(metadata.version, "1.7.0".to_string());
+    assert_eq!(metadata.admin_fee_bps, 5);
     assert_eq!(metadata.state, RunningState::Running);
+
+    // Upgrade to the same code with insurfficient gas.
+    let result = root
+        .call(
+            pool.user_account.account_id.clone(),
+            "upgrade",
+            &EXCHANGE_WASM_BYTES,
+            70_000_000_000_000_u64,
+            0,
+        )
+        .status();
+    assert!(format!("{:?}", result).contains("Not enough gas to complete state migration"));
 
     // Upgrade to the same code migration is skipped.
     root.call(
         pool.user_account.account_id.clone(),
         "upgrade",
         &EXCHANGE_WASM_BYTES,
-        near_sdk_sim::DEFAULT_GAS,
+        100_000_000_000_000_u64,
         0,
     )
     .assert_success();